Problem: error when deserializing OAuth access token for GitHub
They are being serialized as OAuth2AccessToken but attempted to be
read as Strings.

Solution: use String throughout the component. The reason for choosing
String here is because of another discovered flaw: PersistentMap doesn't
play well with Serializables that are not available to every bundle (well,
it simply blows up). That problem requires a separate solution, but for the time
being, we're sticking to the standard library.