Problem: AbstractResourceEmitter stops logging even if no listeners were added
Right now AbstractResourceEmitter has a hardcoded period of 30 seconds to let
interested parties to subscribe for resources, after which it stops logging the
resources. This was done so that it doesn't collect all the resources over
time. The original problem (why does it need to log in the first place?) is
there because we don't know our subscribers upfront as a finite set, they are
self-subscribing.

It will also stop logging even if nobody subscribed within 30 seconds.

Solution: declare listeners as services (not references) and use a whiteboard
pattern so that their respective emitters can be registered and listeners can
be safely connected.

Closes #2