Problem: map:list makes it see the state of the map
It requires a user to lookup bundle IDs in order to look
at the keys and its especially problematic when a good
number of bundles need to be looked into

Solution: make bundleId argument optional.
If not specified, all bundles with PersistentMaps will be
listed.

Also, rename map:list to pmap:list